+++ bin/extract_desktop.dart
@@ -2,17 +2,17 @@ import 'dart:io';
 import 'package:path/path.dart' as p;
 import 'package:webcontent_converter/revision_info.dart';
 
-import 'chromium_helper.dart';
+import 'chrome_helper.dart';
 
 Future<void> main() async {
-  final version = ChromiumInfoConfig.lastVersion;
-  final cachePath = ChromiumInfoConfig.localChromeDirectory;
+  final version = ChromeInfoConfig.lastVersion;
+  final cachePath = ChromeInfoConfig.localChromeDirectory;
   final savePath = Directory.current.path;
 
   print('Using cachePath = $cachePath');
 
   try {
-    final chromeInfo = await ChromiumHelper.justExtractChrome(
+    final chromeInfo = await ChromeHelper.justExtractChrome(
       version: version,
       cachePath: p.join(savePath, cachePath),
     );